Place TestMain() in whitebox test to have the correct package name
