+++ workitem_whitebox_test.go
@@ -1,12 +1,50 @@
 package main
 
 import (
+	"fmt"
+	"os"
 	"testing"
 
+	"github.com/almighty/almighty-core/migration"
+	"github.com/almighty/almighty-core/models"
+	"github.com/almighty/almighty-core/remoteworkitem"
 	"github.com/almighty/almighty-core/resource"
+	"github.com/almighty/almighty-core/transaction"
 	"github.com/goadesign/goa"
+	"github.com/jinzhu/gorm"
 	"github.com/stretchr/testify/assert"
+	"golang.org/x/net/context"
 )
+
+var DB *gorm.DB
+var rwiScheduler *remoteworkitem.Scheduler
+
+func TestMain(m *testing.M) {
+	if _, c := os.LookupEnv(resource.Database); c != false {
+		dbhost := os.Getenv("ALMIGHTY_DB_HOST")
+		if "" == dbhost {
+			panic("The environment variable ALMIGHTY_DB_HOST is not specified or empty.")
+		}
+		var err error
+		DB, err = gorm.Open("postgres", fmt.Sprintf("host=%s user=postgres password=mysecretpassword sslmode=disable", dbhost))
+		if err != nil {
+			panic("failed to connect database: " + err.Error())
+		}
+		defer DB.Close()
+		// Migrate the schema
+		ts := models.NewGormTransactionSupport(DB)
+		witRepo := models.NewWorkItemTypeRepository(ts)
+
+		if err := transaction.Do(ts, func() error {
+			return migration.Perform(context.Background(), ts.TX(), witRepo)
+		}); err != nil {
+			panic(err.Error())
+		}
+		// RemoteWorkItemScheduler now available for all other test cases
+		rwiScheduler = remoteworkitem.NewScheduler(DB)
+	}
+	os.Exit(m.Run())
+}
 
 func TestNewWorkitemController(t *testing.T) {
 	t.Parallel()
